Organize display package & add app shortcuts
