+++ laser_setup/display/__init__.py
@@ -1,4 +1,1 @@
-"""This module includes all GUI elements for the program.
-"""
-from .experiment_window import ExperimentWindow, SequenceWindow  # noqa: F401
-from .main_window import MainWindow, display_window  # noqa: F401
+from .Qt import make_app
