+++ laser_setup/display/widgets/__init__.py
@@ -1,0 +1,3 @@
+from .log_widget import LogsWidget, LogWidget
+from .sqlite_widget import SQLiteWidget
+from .text_widget import TextWidget
