+++ laser_setup/display/widgets/camera_widget.py
